// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/access-user-detail.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface AccessUserDetail {
  createdAtUtc: string;
  displayName: string | null;
  email: string;
  id: number;
  isOwner: boolean;
  normalizedEmail: string;
  permissionCodes: Array<string>;
  status: string;
  sub: string;
  title: string | null;
  updatedAtUtc: string;
  userName: string | null;
  version: number;
}